requirements.txt
pyTelegramBotAPI==4.22.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
